refactor(md): artifact action modals
