+++ app/scripts/modules/core/src/managed/utils/defaults.ts
@@ -1,1 +1,2 @@
 export const TOOLTIP_DELAY = 300;
+export const MODAL_MAX_WIDTH = 750;
